Flush pending handlers between each include.
In particular, run 'apt-get update' right after configured APT, and
restart daemon right after configured them.

The advantage being that if ansible crashes in some "task", the earlier
would already be restarted if neeeded. (This may not happen in the next
run since the configuration should already be up to date.)