Fix Amavis' Policy Banks.
It turns out that in a policy bank, a *_by_ccat doesn't replace the
default but is merely merged into the default (if the keys overlap,
those in the bank take precedence of course).  Hence it's pointless to
use CC_CATCHALL in a bank unless all the other keys have been
overridden, for instance.

Also, treat unchecked (eg, encrypted) mails as clean in the OUTGOING
Policy Bank.