Fix $smtpd_sender_restrictions.
On the MDA the domain is our 'mda.fripost.org', there is no need to
perform an extra DNS lookup.

The MSA does not perform local or virtual delivery, but relays
everything to the outgoing SMTP proxy.

On the MX, there is no need to check for recipient validity as we are
the final destination; but unsure that the RCPT TO address is a valid
recipient before doing the greylisting.