Remove reject_unknown_sender_domain from the MDA and outgoing SMTP.
We already removed it from the MX:es (see 32e605d4); we need to remove
it from the MDA and outgoing SMTP as well, otherwise mails could bounce
or get stuck in the middle (the're rejected with 450: deferred by
default).

However we can keep the restriction on the entry points (MSA and
webmail).